requirements.txt
streamlit==1.39.0
Pillow==10.4.0
ultralytics==8.3.27
pywhatkit==5.2
torch
torchvision
torchaudio